requirements.txt
python-dotenv
pandas
numpy
pymongo
scikit-learn
matplotlib
seaborn
certifi
pymongo[srv]
PyYAML


